requirements.txt
streamlit
pytesseract
pillow
pypdfium2
transformers
torch
sentencepiece
pyttsx3
pydub
numpy
python-dotenv
